validaciones - fixes: ventas

+++ productos/src/main/java/com/apimicroservice/demo/dto/ProductDTO.java
@@ -2,7 +2,7 @@ package com.apimicroservice.demo.dto;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
-
+//
 public record ProductDTO(
     Long id,
 
@@ -12,7 +12,7 @@ public record ProductDTO(
     @Min(value = 1, message = "El precio no puede ser negativo o cero")
     Double price,
     @Min(value = 1, message = "El stock no puede ser negativo o cero")
-    Integer stock,
+    Double stock,
     @Min(value = 1, message = "El stock mínimo no puede ser negativo o cero")
     Integer minStock
 ) {
