+++ productos/src/main/java/com/apimicroservice/demo/model/Product.java
@@ -25,6 +25,6 @@ public class Product {
     private String name;
     private String description;
     private Double price;
-    private Integer stock;
+    private Double stock;
     private Integer minStock;
 }
