+++ ventas/src/main/java/com/apimicroservice/demo/dto/ProductoDTO.java
@@ -4,7 +4,7 @@ public record ProductoDTO(
     Long id,
     String name,
     Double precio,
-    Integer stock
+    Double stock
 ) {
 
 }
